Progress 11: Save Data, Reformatting Code, History Database Implementations

+++ src/data_read/sql/setup_table.sql
@@ -3,5 +3,5 @@ CREATE TABLE IF NOT EXISTS weatherHistory (
     date TEXT,
     location TEXT,
     timestamp TEXT,
-    snapshot_temperature SMALLINT,
+    snapshot_temperature REAL
 )
